SupplyChain-AI/requirements.txt
# Backend
Flask==2.2.2
LangGraph==1.0.0  # Replace with the actual version you use
gunicorn==20.1.0

# LLM and NLP
transformers==4.27.1  # Hugging Face library
torch==1.10.0
torchvision==0.11.0

# Database
mysql-connector-python==8.0.33  # For MySQL
psycopg2-binary==2.9.7  # For PostgreSQL (if using PostgreSQL)

# Frontend API Calls
flask-cors==3.0.10  # CORS support for React frontend

# Data Manipulation
pandas==1.3.5
numpy==1.21.6

# Miscellaneous
tqdm==4.47.0
typing-extensions==3.7.4
Pillow==7.2.0
matplotlib==3.2.2
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-community==0.0.16
langchain-core==0.1.31
langchain-experimental==0.0.47
langchain-groq==0.0.1
azure-cognitiveservices-speech==1.34.0
requests==2.31.0
chromadb==0.4.22
python-dotenv==1.0.0
groq==0.9.0
sentence-transformers==2.2.2
torch==2.1.0